Adding key expiration
Allows the configuration of key expiration based on
x days. Keys will only be valid for that given time
and are removed afterwards. Users are given advanced
notice via mail to replace their keys.

The Database scheme changes and public keys are not
removed anymore. From now on, keys are set as not
active and the fingerprint is unique to prevent users
from reusing their keys.