Add functionality for normal users to view the full data of /servers.json
This patch provides the functionality of allowing certain accounts to
view the full output of /servers.json, including server administrators.

Previously, the only users which could view this data in full were
sysadmins, which does not comply with the least privilege principle of
security (they are able to make changes to keys, while only needing to
view read-only data).